Add .gitignore and new feature Page1 in WebKit project
The commit introduces a .gitignore file which ignores various types of files for different environments and tools, helping to keep the repository clean. Also, a new feature, `Page1`, has been added to the `WebKit` project. This feature includes the associated Razor page, model and updates to the project file and solution.